requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.4.0
paynow>=1.0.7
psycopg2-binary>=2.9.9  # PostgreSQL
django-cors-headers>=4.4.0
gunicorn==20.1.0
django-allauth
whitenoise
dj-database-url
python-dotenv
Pillow
python-decouple
cloudinary 
django-cloudinary-storage
